Exercise2.c

+++ Data Structure & Algorithm/Chapter_2__C_ঝালাই/2.3/Exercise2.c
@@ -7,8 +7,8 @@ int main()
     int n;
     scanf("%d", &n);
 
-    long long int sum = n * (n+1) * ((n+2)/6);
-    printf("%lld", sum);
+    double sum = n * (n+1) * ((n+2)/6);
+    printf("%lf", sum);
 
     return 0;
 }
